Alpha: Build drive support in (test fix)
Dracut doesn't play ball when drive are built as modules, this change builds them in like Fedora does to get around such issues.

This is a workaround fix, the real fix will be to add a Gentoo config on top that saves us from doing this manually on ever rebase.

Signed-off-by: Ian Jordan <immoloism@gmail.com>